replace conditionals with polymorphism
What exactly happens when newParserConfigBuilder(NilPolicy) is run
depends only on NilPolicy, therefore make it a method of NilPolicy.

Also, remove some (probably premature) "optimization" that cached
instance of the Guava*Factory classes in static finals so as to be
able to share instances. On further thought, I don't expect the
creation of parser configurations to occur so often that it would
benefit.

This simplifies the implementation.